+++ x/gamm/pool-models/balancer/pool.go
@@ -769,7 +769,7 @@ func (p *Pool) CalcJoinPoolShares(ctx sdk.Context, tokensIn sdk.Coins, spreadFac
 	newTotalShares := totalShares.Add(numShares)
 
 	// 5) Now single asset join each remaining coin.
-	remainingTokensIn := tokensIn.Sub(tokensJoined)
+	remainingTokensIn := tokensIn.Sub(tokensJoined...)
 	newNumSharesFromRemaining, newLiquidityFromRemaining, err := p.calcJoinSingleAssetTokensIn(remainingTokensIn, newTotalShares, poolAssetsByDenom, spreadFactor)
 	if err != nil {
 		return osmomath.ZeroInt(), sdk.NewCoins(), err
@@ -820,7 +820,7 @@ func (p *Pool) CalcJoinPoolNoSwapShares(ctx sdk.Context, tokensIn sdk.Coins, spr
 	}
 
 	// ensure that no more tokens have been joined than is possible with the given `tokensIn`
-	tokensJoined = tokensIn.Sub(remainingTokensIn)
+	tokensJoined = tokensIn.Sub(remainingTokensIn...)
 	if tokensJoined.IsAnyGT(tokensIn) {
 		return osmomath.ZeroInt(), sdk.NewCoins(), errors.New("an error has occurred, more coins joined than token In")
 	}
@@ -867,7 +867,7 @@ func (p *Pool) ExitPool(ctx sdk.Context, exitingShares osmomath.Int, exitFee osm
 // exitPool exits the pool given exitingCoins and exitingShares.
 // updates the pool's liquidity and totalShares.
 func (p *Pool) exitPool(ctx sdk.Context, exitingCoins sdk.Coins, exitingShares osmomath.Int) error {
-	balances := p.GetTotalPoolLiquidity(ctx).Sub(exitingCoins)
+	balances := p.GetTotalPoolLiquidity(ctx).Sub(exitingCoins...)
 	if err := p.UpdatePoolAssetBalances(balances); err != nil {
 		return err
 	}