+++ x/protorev/keeper/grpc_query.go
@@ -246,3 +246,11 @@ func (q Querier) GetProtoRevPool(c context.Context, req *types.QueryGetProtoRevP
 
 	return &types.QueryGetProtoRevPoolResponse{PoolId: poolId}, nil
 }
+
+// GetAllProtocolRevenue queries all types of protocol revenue (txfees, taker fees, and cyclic arbitrage profits)
+func (q Querier) GetAllProtocolRevenue(c context.Context, req *types.QueryGetAllProtocolRevenueRequest) (*types.QueryGetAllProtocolRevenueResponse, error) {
+	ctx := sdk.UnwrapSDKContext(c)
+	allProtocolRevenue := q.Keeper.GetAllProtocolRevenue(ctx)
+
+	return &types.QueryGetAllProtocolRevenueResponse{AllProtocolRevenue: allProtocolRevenue}, nil
+}