+++ x/tokenfactory/types/msgs_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/osmosis-labs/osmosis/v20/app/apptesting"
 	"github.com/osmosis-labs/osmosis/v20/x/tokenfactory/types"
 
+	"github.com/cometbft/cometbft/crypto/ed25519"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/tendermint/tendermint/crypto/ed25519"
 )
 
 // // Test authz serialize and de-serializes for tokenfactory msg.